+++ src/node2/reference_state.h
@@ -1,5 +1,6 @@
 #ifndef REFERENCE_STATE_H
 #define REFERENCE_STATE_H
+#include <avr/io.h>
 
 typedef struct Reference_State {
     int servo_reference;
@@ -12,13 +13,13 @@ int Get_servo_reference(void);
 
 void Set_servo_reference(int servo_pos);
 
-int Get_motor_reference(void);
+int16_t Get_motor_reference(void);
 
-void Set_motor_reference(int servo_pos);
+void Set_motor_reference(int16_t servo_pos);
 
-int Get_motor_pos();
+int16_t Get_motor_pos();
 
-void Set_motor_pos(int val);
+void Set_motor_pos(int16_t val);
 
 void Set_ir_value(int ir_value);
